requirements.txt
pandas
numpy
google-play-scraper
python-dotenv
tqdm
matplotlib
seaborn
langdetect==1.0.9
transformers>=4.40.0
torch>=2.0.0
scikit-learn
nltk
gensim
psycopg2-binary>=2.9




